requirements.txt
certifi==2023.11.17
langchain==0.0.343
openai==1.3.6
python-dotenv==1.0.0
